registerLibrary change
