+++ src/app/page.tsx
@@ -2,6 +2,7 @@
 
 import Link from 'next/link';
 import { useState } from 'react';
+import Image from 'next/image';
 
 export default function HomePage() {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
@@ -13,7 +14,10 @@ export default function HomePage() {
         <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
           <div className="flex justify-between h-16">
             <div className="flex items-center">
-              <img src="ezhandlelogo.png" alt="projectlogo"  />
+              <Image
+                src="/assets/ezhandlelogo.png" alt="projectlogo" width={40} height={40} className="h-8 w-8 mr-2 rounded-full"
+                onClick={() => setIsMenuOpen(!isMenuOpen)}  
+              />
               <h1 className="text-2xl font-bold text-blue-600 dark:text-blue-400">EZhandle</h1>
             </div>
             <div className="hidden md:flex items-center space-x-4">
